move some resolvers into their own files for cleanup
